pyproject.toml
[project]
name = "ttrsscli"
version = "0.1.0"
authors = [
  { name="Peter Reuterås", email="peter@reuteras.net" },
]
description = "Textual tool to access a Tiny Tiny RSS instance to read articles."
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "cleanurl>=0.1.15",
    "httpx>=0.28.1",
    "markdownify>=0.14.1",
    "pillow>=11.1.0",
    "readwise-api>=0.3.0",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "rich-pixels>=3.0.1",
    "textual-serve>=1.1.1",
    "textual[syntax]>=1.0.0",
    "toml>=0.10.2",
    "ttrss-python",
]
[project.scripts]
ttrsscli = "ttrsscli.main:main"
ttrsscli-web = "ttrsscli.main:main_web"

[project.urls]
Homepage = "https://github.com/reuteras/ttrsscli"
Issues = "https://github.com/reuteras/ttrsscli/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pylint>=3.3.4",
    "pytest>=8.3.4",
    "ruff>=0.9.5",
    "textual-dev>=1.7.0",
]

[tool.ruff]
target-version = "py311"
lint.select = [
    "PL",   # Pylint
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "RUF",  # Ruff-specific
]
lint.ignore = [
    "E501",  # line too long, handled by black
]
fix = true

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.uv.sources]
ttrss-python = { git = "https://github.com/reuteras/ttrss-python.git" }

[tool.ruff.lint.pydocstyle]
convention="google"
